Extract parameterized test
